refactor(ShowTVDBCreator): break 536-line component into hook + parts
- useShowTVDBCreator hook owns search/select/create state and the create-show
  API call
- ShowSearchView, ShowDetailsForm, EpisodesPreviewList extracted
- Reuses the shared FormField primitives (extended with optional labelStyle to
  preserve this modal's var(--text) labels); keeps its own plain overlay (not a
  Radix/motion modal) to avoid any visual change
- Folder with index.tsx orchestrator; RTPImporter import unchanged

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
